bist/requirements.txt
playwright>=1.49.0
anthropic>=0.43.0
click>=8.1.7
rich>=13.7.0
Pillow>=10.0.0
aiohttp>=3.9.0
